Stop type casting int's, 32 bit appears to hate it
git-svn-id: https://svn.code.sf.net/p/freeghost/code/trunk@5742 71f96598-fa45-0410-b640-bcd6f8691b32

+++ packages/web/commons/init.php
@@ -132,7 +132,7 @@ $DB = FOGCore::getClass('DatabaseManager')->establish()->getDB();
 $FOGCore::setSessionEnv();
 /** $TimeZone the timezone setter */
 $TimeZone = $_SESSION['TimeZone'];
-$currentUser = FOGCore::getClass('User',(int)(isset($_SESSION['FOG_USER']) ? $_SESSION['FOG_USER'] : 0));
+$currentUser = FOGCore::getClass('User',(isset($_SESSION['FOG_USER']) ? $_SESSION['FOG_USER'] : 0));
 $HookManager = FOGCore::getClass('HookManager');
 $HookManager->load();
 $EventManager = FOGCore::getClass('EventManager');
