+++ packages/web/lib/service/multicastmanager.class.php
@@ -7,7 +7,7 @@ class MulticastManager extends FOGService {
         static::$log = sprintf('%s%s',self::$logpath ? self::$logpath : '/opt/fog/log/',$log ? $log : 'multicast.log');
         if (file_exists(static::$log)) unlink(static::$log);
         static::$dev = $dev ? $dev : '/dev/tty2';
-        static::$zzz = (int)($zzz ? $zzz : 10);
+        static::$zzz = ($zzz ? $zzz : 10);
     }
     private function isMCTaskNew($KnownTasks, $id) {
         foreach((array)$KnownTasks AS $i => &$Known) $output[] = $Known->getID();
@@ -120,7 +120,7 @@ class MulticastManager extends FOGService {
                 self::outall($e->getMessage());
             }
             self::out(' +---------------------------------------------------------',static::$dev);
-            $tmpTime = (int)self::getSetting(self::$sleeptime);
+            $tmpTime = self::getSetting(self::$sleeptime);
             if (static::$zzz != $tmpTime) {
                 static::$zzz = $tmpTime ? $tmpTime : 10;
                 self::outall(sprintf(" | Sleep time has changed to %s seconds",static::$zzz));
